@@ -5,20 +5,20 @@ description: A web app that uses AngularDart Components
 # author: Thai Huynh <thaihuynhxyz@gmail.com>
 
 environment:
-  sdk: '>=2.2.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  equatable: ^0.2.0
-  flutter_bloc: ^0.21.0
-  intl: ^0.15.8
+  equatable: ^2.0.3
+  flutter_bloc: ^8.0.1
+  intl: ^0.17.0
   
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^4.0.0
-  build_runner: ^1.6.7
+  mockito: ^5.2.0
+  build_runner: ^2.1.11
 
 flutter:
   fonts:
