@@ -3,6 +3,6 @@ import 'package:flutter/material.dart';
 class Choice {
   const Choice({this.title, this.icon});
 
-  final String title;
-  final IconData icon;
+  final String? title;
+  final IconData? icon;
 }
