@@ -2,11 +2,11 @@ class Lunar {
   int day;
   int month;
   int year;
-  String monthString;
-  String canChiDay;
-  String canChiMonth;
-  String canchiYear;
-  List<String> hours;
+  String? monthString;
+  String? canChiDay;
+  String? canChiMonth;
+  String? canchiYear;
+  List<String>? hours;
 
   Lunar(this.day, this.month, this.year, this.monthString, this.canChiDay,
       this.canChiMonth, this.canchiYear, this.hours);
