@@ -1,9 +1,10 @@
+import 'package:equatable/equatable.dart';
 import 'package:licham/lunar.dart';
 import 'package:meta/meta.dart';
-import 'package:equatable/equatable.dart';
 
 abstract class MainState extends Equatable {
-  MainState([List props = const []]) : super(props);
+  @override
+  List<Object?> get props => const [];
 }
 
 class MainUninitialized extends MainState {
@@ -14,19 +15,24 @@ class MainUninitialized extends MainState {
 class MainFailure extends MainState {
   final String error;
 
-  MainFailure({@required this.error}) : super([error]);
+  MainFailure({required this.error});
 
   @override
   String toString() => 'LoginFailure { error: $error }';
+
+  @override
+  List<Object> get props => [error];
 }
 
 class DateUpdate extends MainState {
-  final DateTime solar;
-  final Lunar lunar;
+  final DateTime? solar;
+  final Lunar? lunar;
 
-  DateUpdate({@required this.solar, @required this.lunar})
-      : super([solar, lunar]);
+  DateUpdate({required this.solar, required this.lunar});
 
   @override
   String toString() => 'DateUpdate { solar: $solar, lunar: $lunar }';
+
+  @override
+  List<Object?> get props => [solar, lunar];
 }
