@@ -3,14 +3,14 @@ import 'package:licham/main_bloc.dart';
 import 'package:licham/main_state.dart';
 
 void main() {
-  MainBloc mainBloc;
+  late MainBloc mainBloc;
 
   setUp(() {
     mainBloc = MainBloc(DateTime.now());
   });
 
   test('initial state is correct', () {
-    expect(mainBloc.initialState, MainUninitialized());
+    expect(mainBloc.state, MainUninitialized());
   });
 
   test('dispose does not emit new states', () {
@@ -18,6 +18,6 @@ void main() {
       mainBloc.state,
       emitsInOrder([MainUninitialized(), emitsDone]),
     );
-    mainBloc.dispose();
+    mainBloc.close();
   });
 }
