@@ -1,4 +1,3 @@
-import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:intl/intl.dart';
@@ -6,34 +5,12 @@ import 'package:licham/main_bloc.dart';
 import 'package:licham/main_event.dart';
 import 'package:licham/main_view.dart';
 
-class SimpleBlocDelegate extends BlocDelegate {
-  @override
-  void onEvent(Bloc bloc, Object event) {
-    super.onEvent(bloc, event);
-    print(event);
-  }
-
-  @override
-  void onTransition(Bloc bloc, Transition transition) {
-    super.onTransition(bloc, transition);
-    print(transition);
-  }
-
-  @override
-  void onError(Bloc bloc, Object error, StackTrace stacktrace) {
-    super.onError(bloc, error, stacktrace);
-    print(error);
-  }
-}
-
 void main() {
-  BlocSupervisor.delegate = SimpleBlocDelegate();
   DateTime date = _getDate(Uri.base.queryParameters);
   runApp(
     BlocProvider<MainBloc>(
-      builder: (context) {
-        return MainBloc(date)
-          ..dispatch(AppStarted());
+      create: (context) {
+        return MainBloc(date)..add(AppStarted());
       },
       child: App(),
     ),
@@ -53,7 +30,9 @@ class App extends StatelessWidget {
 DateTime _getDate(Map<String, String> queryParameters) {
   final date = queryParameters['date'];
   try {
-    return DateFormat("dd-MM-yyyy", "en_US").parse(date);
+    return date != null
+        ? DateFormat("dd-MM-yyyy", "en_US").parse(date)
+        : DateTime.now();
   } on Exception {
     return DateTime.now();
   }
@@ -72,10 +51,8 @@ ThemeData _buildDarkTheme() {
     primaryColor: primaryColor,
     primaryColorDark: const Color(0xFF0050a0),
     primaryColorLight: secondaryColor,
-    buttonColor: primaryColor,
     indicatorColor: Colors.white,
     toggleableActiveColor: const Color(0xFF6997DF),
-    accentColor: secondaryColor,
     canvasColor: const Color(0xFF202124),
     scaffoldBackgroundColor: const Color(0xFF202124),
     backgroundColor: const Color(0xFF202124),
